lobbyspy-speech-retrieval/requirements.txt
boto3==1.34.13
pandas==2.0.3
pinecone-client==2.2.4
requests